Refactor: ProfileSection의 모든 요소들 정리
